packages/common/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "claresys-common"
version = "0.1.0"
description = "Shared utilities for ClaReSys microservices"
requires-python = ">=3.10"
dependencies = [
  "python-jose",
  "pydantic",
  "fastapi"
]
